fix: fixed duplicate code
